refactor: give the pricing FAQ the homepage accordion
The pricing page had its own FAQ built on a $state toggle and plain buttons,
with no animation and a different visual treatment to the homepage one.

Extract the homepage FAQ into MarketingFaq, taking its questions, its details
grouping name, and an optional title as props. Both pages now render the same
accordion; the pricing page keeps its own billing questions.

Net 75 fewer lines: the pricing page sheds its FAQ markup, its toggle state,
and its entire style block.
